# openapi.json
{"openapi":"3.1.0","info":{"title":"SocialPulse AI","version":"0.1.0"},"paths":{"/api/health":{"get":{"summary":"Health","operationId":"health_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/brands":{"get":{"summary":"Brands","operationId":"brands_api_brands_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"summary":"Create Brand","operationId":"create_brand_api_brands_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BrandIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/brands/{brand_id}":{"get":{"summary":"Get Brand","operationId":"get_brand_api_brands__brand_id__get","parameters":[{"name":"brand_id","in":"path","required":true,"schema":{"type":"integer","title":"Brand Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/brands/{brand_id}/posts":{"post":{"summary":"Create Post","operationId":"create_post_api_brands__brand_id__posts_post","parameters":[{"name":"brand_id","in":"path","required":true,"schema":{"type":"integer","title":"Brand Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Brand Posts","operationId":"get_brand_posts_api_brands__brand_id__posts_get","parameters":[{"name":"brand_id","in":"path","required":true,"schema":{"type":"integer","title":"Brand Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/brands/{brand_id}/seed":{"post":{"summary":"Seed","operationId":"seed_api_brands__brand_id__seed_post","parameters":[{"name":"brand_id","in":"path","required":true,"schema":{"type":"integer","title":"Brand Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/brands/{brand_id}/analytics":{"get":{"summary":"Get Brand Analytics","operationId":"get_brand_analytics_api_brands__brand_id__analytics_get","parameters":[{"name":"brand_id","in":"path","required":true,"schema":{"type":"integer","title":"Brand Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/brands/{brand_id}/learn":{"post":{"summary":"Learn","operationId":"learn_api_brands__brand_id__learn_post","parameters":[{"name":"brand_id","in":"path","required":true,"schema":{"type":"integer","title":"Brand Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/brands/{brand_id}/recommendations":{"post":{"summary":"Create Recommendations","operationId":"create_recommendations_api_brands__brand_id__recommendations_post","parameters":[{"name":"brand_id","in":"path","required":true,"schema":{"type":"integer","title":"Brand Id"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecommendationRequest","default":{"query":"What should I post tomorrow?"}}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Get Recommendations History","operationId":"get_recommendations_history_api_brands__brand_id__recommendations_get","parameters":[{"name":"brand_id","in":"path","required":true,"schema":{"type":"integer","title":"Brand Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/brands/{brand_id}/recommendations/{rec_id}/action":{"post":{"summary":"Update Recommendation Action","operationId":"update_recommendation_action_api_brands__brand_id__recommendations__rec_id__action_post","parameters":[{"name":"brand_id","in":"path","required":true,"schema":{"type":"integer","title":"Brand Id"}},{"name":"rec_id","in":"path","required":true,"schema":{"type":"integer","title":"Rec Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RecommendationActionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/brands/{brand_id}/recommend":{"post":{"summary":"Recommend","operationId":"recommend_api_brands__brand_id__recommend_post","parameters":[{"name":"brand_id","in":"path","required":true,"schema":{"type":"integer","title":"Brand Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Question"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/brands/{brand_id}/feedback":{"post":{"summary":"Feedback","operationId":"feedback_api_brands__brand_id__feedback_post","parameters":[{"name":"brand_id","in":"path","required":true,"schema":{"type":"integer","title":"Brand Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Feedback"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/brands/{brand_id}/memory":{"get":{"summary":"Memories","operationId":"memories_api_brands__brand_id__memory_get","parameters":[{"name":"brand_id","in":"path","required":true,"schema":{"type":"integer","title":"Brand Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/brands/{brand_id}/memory/status":{"get":{"summary":"Memory Status","operationId":"memory_status_api_brands__brand_id__memory_status_get","parameters":[{"name":"brand_id","in":"path","required":true,"schema":{"type":"integer","title":"Brand Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"BrandIn":{"properties":{"name":{"type":"string","title":"Name"},"industry":{"type":"string","title":"Industry","default":"CafÃ©"},"business_description":{"type":"string","title":"Business Description","default":""},"target_audience":{"type":"string","title":"Target Audience","default":""},"tone":{"type":"string","title":"Tone","default":"Friendly"},"language":{"type":"string","title":"Language","default":"English"},"products_services":{"type":"string","title":"Products Services","default":""},"location":{"type":"string","title":"Location","default":""},"goals":{"type":"string","title":"Goals","default":""},"competitors":{"type":"string","title":"Competitors","default":""},"avoid_topics":{"type":"string","title":"Avoid Topics","default":""}},"type":"object","required":["name"],"title":"BrandIn"},"Feedback":{"properties":{"feedback":{"type":"string","title":"Feedback"}},"type":"object","required":["feedback"],"title":"Feedback"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PostIn":{"properties":{"platform":{"type":"string","title":"Platform","default":"Instagram"},"published_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Published At"},"content_format":{"type":"string","title":"Content Format","default":"Reel"},"category":{"type":"string","title":"Category","default":"Educational"},"topic":{"type":"string","title":"Topic","default":""},"caption":{"type":"string","title":"Caption","default":""},"status":{"type":"string","title":"Status","default":"published"},"impressions":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Impressions"},"reach":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Reach"},"likes":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Likes"},"comments":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Comments"},"shares":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Shares"},"saves":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Saves"},"link_clicks":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Link Clicks"},"follower_growth":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Follower Growth"},"metrics":{"anyOf":[{"$ref":"#/components/schemas/PostMetricsIn"},{"type":"null"}]}},"type":"object","title":"PostIn"},"PostMetricsIn":{"properties":{"impressions":{"type":"integer","title":"Impressions","default":0},"reach":{"type":"integer","title":"Reach","default":0},"likes":{"type":"integer","title":"Likes","default":0},"comments":{"type":"integer","title":"Comments","default":0},"shares":{"type":"integer","title":"Shares","default":0},"saves":{"type":"integer","title":"Saves","default":0},"link_clicks":{"type":"integer","title":"Link Clicks","default":0},"follower_growth":{"type":"integer","title":"Follower Growth","default":0}},"type":"object","title":"PostMetricsIn"},"Question":{"properties":{"question":{"type":"string","title":"Question","default":"What should I post tomorrow?"}},"type":"object","title":"Question"},"RecommendationActionRequest":{"properties":{"action":{"type":"string","title":"Action","default":"approve"},"recommendation_index":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Recommendation Index","default":0},"edited_content":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Edited Content"},"feedback":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Feedback"}},"type":"object","title":"RecommendationActionRequest"},"RecommendationRequest":{"properties":{"query":{"type":"string","title":"Query","default":"What should I post tomorrow?"}},"type":"object","title":"RecommendationRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}